Popover: carefully cancel deferred tasks when hiding/unmounting the component
The `Popover` component uses `_.defer` to delay certain tasks, like reposition on
rerender or focus after show, to the next event loop ticks.

However, in certain edge cases (that happen often enough) the `Popover` can be hidden
or unmounted before these tasks have a chance to run. In such case, they try to work
with a DOM element that no longer exist and can crash.

This patch carefully tracks these deferred tasks and cancels them when the `Popover`
is being hidden or unmounted.