Add check for block already registered
Adds an extra check if the block is already registerd, so there is not
an error if it is. This is needed for coordinating this plugin
deployment on WPCOM, which has a separate plugin registering the blocks.

The same checks were added to that plugin, so it won't make a difference
which plugin gets called first.

These checks can be removed after deployment and retirement of the other
plugin.